% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_numeric}
\alias{as_numeric}
\title{Convert to numeric passing through character for safety}
\usage{
as_numeric(x, ..., require_conversion = NULL)
}
\arguments{
\item{x}{vector}

\item{...}{additional argument to as.character}

\item{require_conversion}{If values are converted to \code{NA}, should
nothing be done (\code{NULL}), a warning be printed (\code{FALSE}), or an
error occur (\code{TRUE})?  (Inputs that are \code{NA} do not trigger the
warning or error.)}
}
\description{
Convert to numeric passing through character for safety
}
\examples{
# factor with weird levels that we don't want to keep
ex <- factor(c(1, 2, 3, 4), levels = c(2, 3, 1, 4)) 
ex

# keeps information about the levels, oh no!
as.numeric(ex) 

# keeps the labelled values
as_numeric(ex)

as_numeric(c("1", "A"), require_conversion=FALSE)
}
\seealso{
Other Numerics: 
\code{\link{unique_non_numerics}()}
}
\concept{Numerics}
