% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AUC_partial.R
\name{AUC_partial}
\alias{AUC_partial}
\title{Calculate partial AUC}
\usage{
AUC_partial(time, conc, range = c(0, max(time)))
}
\arguments{
\item{time}{time column}

\item{conc}{concentration column}

\item{range}{time range for pauc calculation}
}
\description{
Calculate partial AUC
}
\details{
default range is 0 to tmax
is recommended to be used alongside dplyr for ease of calculation
}
\examples{
\dontrun{
df \%>\% group_by(ID) \%<\% summarize(pAUC0-10 = AUC_partial(TIME, DV, c(0,10)))
}
}

