% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{auc_partial_cpp}
\alias{auc_partial_cpp}
\title{calculate partial AUC}
\usage{
auc_partial_cpp(time, dv, range)
}
\arguments{
\item{time}{vector of time values}

\item{dv}{concentration measurements}

\item{range}{vector of min and max value of the partial auc range}
}
\description{
calculate partial AUC
}
