% Generated by roxygen2 (4.0.1): do not edit by hand
\name{base_theme}
\alias{base_theme}
\title{A theme with better default values for pharmacometric plots, especially conc-time}
\usage{
base_theme()
}
\description{
A theme with better default values for pharmacometric plots, especially conc-time
}

