% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{basename_sans_ext}
\alias{basename_sans_ext}
\title{get the basename of a filepath, minus any extensions}
\usage{
basename_sans_ext(.x)
}
\arguments{
\item{.x}{filepath}
}
\description{
get the basename of a filepath, minus any extensions
}
