% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wam.R
\name{calculate_wam}
\alias{calculate_wam}
\title{internal WAM function}
\usage{
calculate_wam(k, p, n, theta, cov)
}
\arguments{
\item{k}{number of covariate thetas}

\item{p}{total number of parameters in model (theta, omega, sigma)}

\item{n}{number of observations}

\item{theta}{vector of covariate parameter values}

\item{cov}{variance-covariance matrix for covariate parameters}
}
\value{
data.frame
}
\description{
internal WAM function
}
\details{
returns model rank, the LRT, SBC, which covariates selected
for internal use with wam()
}
