% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_names.R
\name{capitalize_names}
\alias{capitalize_names}
\title{capitalize all names for a dataframe}
\usage{
capitalize_names(df)
}
\arguments{
\item{df}{data frame to capitalize names}
}
\description{
capitalize all names for a dataframe
}
\details{
is a simple wrapper function to reduce typing and more easily pass data
as it is read from a file
}
\examples{
names(Theoph)
cTheoph <- capitalize_names(Theoph)
names(cTheoph)
\dontrun{
df <- capitalize_names(df)
# make sure all names are capitalized as a file is read
df <- capitalize_names(read.csv(...))
}
}

