% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/groups.R
\name{capture_groups}
\alias{capture_groups}
\title{wrapper to capture groups from a grouped data frame}
\usage{
capture_groups(df)
}
\arguments{
\item{data}{frame to determine groups}
}
\description{
wrapper to capture groups from a grouped data frame
}
\examples{
\dontrun{
gTheoph <- dplyr::group_by(Theoph, Subject)
capture_groups(gTheoph)
}
}

