% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunking.R
\name{ids_per_plot}
\alias{ids_per_plot}
\alias{chunk_grp}
\alias{chunk}
\alias{chunk_list}
\alias{chunk_grp_list}
\title{split IDs into groups to use for subsequent plotting}
\usage{
ids_per_plot(id, id_per_plot = 9)

chunk_grp(.x, .nchunk)

chunk(.x, .nchunk)

chunk_list(.x, .nchunk)

chunk_grp_list(.x, .nchunk)
}
\arguments{
\item{id}{vector of ids (eg id column)}

\item{id_per_plot}{number of ids per plot. Default to 9}

\item{.x}{vector of values}

\item{.nchunk}{number of chunks to identify}
}
\description{
split IDs into groups to use for subsequent plotting

chunk
}
\details{
works very well with hadley wickham's purrr package to create a column
to split on then subsequently plot, see \code{vignette("Multiplot")} for details

chunk by group, unique values, and return as a vector or a list with elememts
}
