% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean_nsub.R
\name{clean_nsub}
\alias{clean_nsub}
\title{cleans nsub output}
\usage{
clean_nsub(path, title = "TABLE", header = "DV", overwrite = T)
}
\arguments{
\item{path}{path to file}

\item{title}{keyword to determine title rows (eg. TABLE)}

\item{header}{keyword to determine header rows (eg DV)}

\item{overwrite}{whether to overwrite the original file. If FALSE, will add a '_clean'
to existing fle name}
}
\description{
cleans nsub output
}

