% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cols_to_factor}
\alias{cols_to_factor}
\title{convert column to type factor}
\usage{
cols_to_factor(df, col_names)
}
\arguments{
\item{df}{dataframe}

\item{col_names}{vector of column names or indices to convert to factor}
}
\description{
convert column to type factor
}
\examples{
\dontrun{
cols_to_factor(df, c("DOSE", "TRT")) # will convert dose and TRT columns to factor
}
}

