% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cols_to_numeric.R
\name{cols_to_numeric}
\alias{cols_to_numeric}
\title{convert column to type numeric}
\usage{
cols_to_numeric(df, col_names)
}
\arguments{
\item{df}{dataframe}

\item{col_names}{vector of column names or indices to convert to numeric}
}
\description{
convert column to type numeric
}
\examples{
\dontrun{
df <- cols_to_numeric(df, c("DOSE", "TRT")) # will convert dose and TRT columns to numeric
}
}

