% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_covariance.R
\name{cor_to_cov}
\alias{cor_to_cov}
\title{correlation to covariance matrix}
\usage{
cor_to_cov(.cor_mat, .sd)
}
\arguments{
\item{.cor_mat}{correlation matrix}

\item{.sd}{vector of standard deviations}
}
\description{
correlation to covariance matrix
}
