% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_covariance.R
\name{cov_to_cor}
\alias{cov_to_cor}
\title{covariance to correlation}
\usage{
cov_to_cor(.cov_mat)
}
\arguments{
\item{.cov_mat}{covariance matrix}
}
\value{
list of unit correlation matrix and vector of standard deviations
}
\description{
covariance to correlation
}
