% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-subject-check.R
\name{df_subject_check}
\alias{df_subject_check}
\title{Subject data checks}
\usage{
df_subject_check(.df, .subject_col = NULL, .col_check = NULL)
}
\arguments{
\item{.df}{Data.frame to perform checks on}

\item{.subject_col}{Character column name of subject identifier. (default to ID or USUBJID)}

\item{.col_check}{Character column name of variable to check if subject only has 1 distinct value of}
}
\description{
Use this function to perform quick checks on a data.frame to determine if only one row or
values exists for each subject within a column. If `.col_check` is `NULL` (which is the default) 
then the function will check if the data.frame only has 1 row per subject.

If a variable is provided to `.col_check` then the function will check if each subject only has 
1 distinct value within that column.
}
\examples{
library(dplyr)
Theoph2 <- Theoph \%>\% dplyr::group_by(Subject) \%>\% dplyr::slice(1) \%>\% dplyr::ungroup()

# Check if each subject has only one row
df_subject_check(Theoph2, .subject_col = "Subject")

# Check if each subject has only one distinct Dose value
df_subject_check(Theoph2, .subject_col = "Subject", .col_check = "Dose")

}
