% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_factors.R
\name{drop_factors}
\alias{drop_factors}
\title{Drop factor level if not present in data}
\usage{
drop_factors(.df)
}
\arguments{
\item{.df}{data.frame to be modified}
}
\description{
Returns a data.frame with factor levels removed if they do
not exist in the data. Columns that are not factors or all
levels are present will not be changed.
}
\examples{
.df <- data.frame(
   X = factor(c("a", "b"), levels = c("a", "b", "c")),
   Y = 1:2)

drop_factors(.df)

}
