% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_curve_stripping.R
\name{execute_curve_stripping}
\alias{execute_curve_stripping}
\title{curve strip a dataframe with multiple IDs}
\usage{
execute_curve_stripping(df, TIME_name = "TIME", DV_name = "CONC",
  DOSE_name = "DOSE", number_terminal_points, ...)
}
\arguments{
\item{df}{dataframe to curve strip}

\item{TIME_name}{column name for time column}

\item{DV_name}{column name for dependent variable column eg. conc}

\item{DOSE_name}{column name for dose column}

\item{number_terminal_points}{number of data points in terminal phase}

\item{...}{additional args to be passed to strip_curves}
}
\description{
curve strip a dataframe with multiple IDs
}
\details{
calculates the geometric mean by dose group then curve strips each mean
expects 3 columns will be in dataset, the independent variable (eg time), 
dependent variable (eg concentration), and a dose column with the dose given for each ID in all rows
}
\examples{
\dontrun{
execute_curve_stripping(dat_IV, number_terminal_points =9)
execute_curve_stripping(dat_IV, "time", "CObs", "DOSE", number_terminal_points=6)
}
}
