% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resampling_functions.R
\name{get_key}
\alias{get_key}
\title{find unique values for key}
\usage{
get_key(df, key_cols)
}
\arguments{
\item{df}{data frame}

\item{key_cols}{vector of column names}
}
\description{
find unique values for key
}

