% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{ghere}
\alias{ghere}
\title{Glue variables into here::here() path}
\usage{
ghere(..., .envir = rlang::caller_env())
}
\arguments{
\item{...}{Passed to [here::here()].}

\item{.envir}{Passed to [glue::glue()].}
}
\value{
The path with class character after string interpolation.
}
\description{
Glue variables into here::here() path
}
\examples{
\dontrun{
 run <- 12345
 ghere("deliv", "figure", "{run}")
}

}
