% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_life.R
\name{half_life}
\alias{half_life}
\title{Estimate elimination half-life from 2-compartment model}
\usage{
half_life(cl, vc, vp, q)
}
\arguments{
\item{cl}{clearance estimate (L/hr)}

\item{vc}{central volume estimate (L)}

\item{vp}{peripheral volume estimate (L)}

\item{q}{intercompartmental clearance estimate (L/hr)}
}
\description{
Result provided in hours.
}
\examples{
half_life(cl = 1, vc = 51, vp = 13, q = 1)

}
\author{
Samuel P Callisto, PhD
}
