% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_ext.R
\name{has_ext}
\alias{has_ext}
\title{check for if the passed name has a file extension}
\usage{
has_ext(name, ext, match_case = TRUE)
}
\arguments{
\item{name}{string name to check for extension}

\item{ext}{string of extension to check}

\item{match_case}{logical whether to match the case when checking extension.
defaults to TRUE}
}
\description{
check for if the passed name has a file extension
}
\details{
This is not a particularly robust checker, but serves its purpose
}
\examples{
\dontrun{
has_ext("test.rmd", ".rmd") #TRUE
has_ext("test.Rmd", ".rmd", match_case=F) #TRUE
has_ext("testrmd", ".rmd") #FALSE
}
}
