% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{hash}
\alias{hash}
\title{Add hash character to start of each line}
\usage{
hash(x, char = "#")
}
\arguments{
\item{x}{string}

\item{char}{hash character string}
}
\description{
The given expression will be returned, with a hash character at the start of each line.
}
