% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\name{ht}
\alias{ht}
\title{Display top and bottom of data frame}
\usage{
ht(.df, .n = 4)
}
\arguments{
\item{.df}{data.frame to display}

\item{.n}{number of rows to display from each end of data frame}
}
\description{
Display top and bottom of data frame
}
\examples{

# View the top 4 and bottom 4 rows of a dataframe
ht(Theoph, 4)

}
