% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_sas.R
\name{inspect_sas}
\alias{inspect_sas}
\title{Function to inspect all SAS files in a given directory
either displays files one at a time (default), or opens all files at once in separate windows}
\usage{
inspect_sas(.dir, .cycle = TRUE)
}
\arguments{
\item{.dir}{directory containing SAS files}

\item{.cycle}{if TRUE, waits for user input to cycle through. if FALSE, opens all files simultaneously.}
}
\description{
Function to inspect all SAS files in a given directory
either displays files one at a time (default), or opens all files at once in separate windows
}
