% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-distinct.R
\name{is_distinct}
\alias{is_distinct}
\title{Check if a group of columns are distinct}
\usage{
is_distinct(.df, .cols)
}
\arguments{
\item{.df}{A data.frame}

\item{.cols}{List of column names to perform check on}
}
\description{
Provide a data.frame along with any number of column names that exist in the data and 
see if all rows are distinct. Can be useful when checking for duplicates.
}
\examples{
duplicate_df <- 
  dplyr::tibble(
    ID = c(1, 1, 2, 2),
    EVID = c(0, 1, 1, 1),
    TIME = c(-0.5, 1, 1, 1))

is_distinct(.df = duplicate_df, .cols = c("EVID", "TIME"))

}
