% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jprint.R
\name{jprint}
\alias{jprint}
\title{print multiple values joined together}
\usage{
jprint(..., sep = " ")
}
\arguments{
\item{...}{variadic parameter to of objects to combine}

\item{sep}{separator value}
}
\description{
print multiple values joined together
}
\details{
print can only take one value to be printed, this function j(oined)print
 provides a wrapper to obviate the need to do print(paste0(...)) type work-arounds
}
\examples{
result <- "world"
jprint("hello", result)
}

