% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_names.R
\name{lowercase_names}
\alias{lowercase_names}
\title{lowercase all names for a dataframe}
\usage{
lowercase_names(df)
}
\arguments{
\item{df}{data frame to lowercase names}
}
\description{
lowercase all names for a dataframe
}
\details{
is a simple wrapper function to reduce typing and more easily pass data
as it is read from a file
}
\examples{
\dontrun{
df <- lowercase_names(df)
# make sure all names are lowered as a file is read
df <- lowercase_names(read.csv(...))
# or via dplyr pipe syntax
df <- read.csv(...) \%>\% lowercase_names
}
}

