% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_through}
\alias{max_through}
\title{Give the max value up to that point}
\usage{
max_through(x)
}
\arguments{
\item{x}{A numeric vector of values}
}
\description{
Give the max value up to that point
}
\details{
useful for seeing how a current measurement compares to the maximum
value up to that point
}
\examples{
max_through(c(4, 3, 3, 2, 5, 1))
[1] 4 4 4 4 5 5

max_through(c(NA, 2, 1, 4, 2))
[1] NA  2  2  4  4

}
