% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgmisc.R
\docType{package}
\name{mrgmisc-package}
\alias{mrgmisc}
\alias{mrgmisc-package}
\title{mrgmisc: MRG Helpers}
\description{
Metrum Research Group Helper Functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/metrumresearchgroup/mrgmisc}
  \item \url{https://metrumresearchgroup.github.io/mrgmisc}
  \item Report bugs at \url{https://github.com/metrumresearchgroup/mrgmisc/issues}
}

}
\author{
\strong{Maintainer}: Michael McDermott \email{michaelm@metrumrg.com}

Authors:
\itemize{
  \item Devin Pastoor \email{devin.pastoor@gmail.com}
  \item Kyle Baron \email{kyleb@metrumrg.com}
  \item Grace O'Brien \email{graceo@metrumrg.com}
  \item Kyle Barrett \email{barrettk@metrumrg.com}
  \item Tim Bergsma \email{timb@metrumrg.com}
}

}
\keyword{internal}
