% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasum.R
\name{nasum}
\alias{nasum}
\title{Display a count of NAs by column in a data.frame}
\usage{
nasum(.df)
}
\arguments{
\item{.df}{the count of NAs will be calculated for each column in this data 
set, with any grouping variables retained}
}
\description{
\code{nasum} Count the number of \code{NA} values in each element of a list-like object.
}
\details{
Arguments are passed to \code{simplify}, along with a function 
that sums instances of \code{NA}.
}
\examples{
nasum(Theoph)

}
\author{
Tim Waterhouse
}
