% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NCA.R
\name{NCA}
\alias{NCA}
\title{calculate basic NCA parameters t1/2 Cl, Cmax etc}
\usage{
NCA(TIME = "TIME", DV = "DV", DOSE = "DOSE", last_times = c(3, 4, 5))
}
\arguments{
\item{TIME}{vector of times}

\item{DV}{vector of observations (concentrations)}

\item{DOSE}{vector or single value of Dose given}

\item{last_times}{vector of numbers of time points to evaluate for
AUCinf extrapolation default 3-5}
}
\description{
calculate basic NCA parameters t1/2 Cl, Cmax etc
}

