% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/new_rmd.R
\name{new_report}
\alias{new_report}
\title{a better new rmd template for reports}
\usage{
new_report(report)
}
\arguments{
\item{file}{file name and (optionally) path to subdirectory}

\item{title}{Title of rmd document}

\item{author}{Author of document}

\item{date}{Date}

\item{template}{template file to use for file creation}
}
\description{
a better new rmd template for reports
}
\examples{
\dontrun{
new_report("lab-notebook/Report.Rmd")
new_report("Report.Rmd")
}
}

