% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/new_rmd.R
\name{new_rmd}
\alias{new_rmd}
\title{a better new rmd template}
\usage{
new_rmd(file, title = NULL, author = NULL, date = format(Sys.time(),
  "\%b \%d, \%Y"), template = "basic.Rmd")
}
\arguments{
\item{file}{file name and (optionally) path to subdirectory}

\item{title.}{Title of rmd document}

\item{author.}{Author of document}

\item{date.}{Date}

\item{template.}{template file to use for file creation}
}
\description{
a better new rmd template
}
\examples{
\dontrun{
new_rmd("Report.Rmd")
new_rmd("templates/Report.Rmd", "Example Report", "Devin")
}
}

