% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsub.R
\name{nsub}
\alias{nsub}
\title{Count number of IDs in a NONMEM data set}
\usage{
nsub(.df, .subject_col)
}
\arguments{
\item{.df}{data.frame containing subject column}

\item{.subject_col}{column name that will be used to uniquely identify IDs}
}
\description{
Count number of IDs in a NONMEM data set
}
\examples{

nsub(.df = Theoph, .subject_col = "Subject")

}
\author{
Samuel P Callisto, PhD
}
