% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_time.R
\name{numeric_time}
\alias{numeric_time}
\title{Convert clock time to decimal equivalent}
\usage{
numeric_time(.col)
}
\arguments{
\item{.col}{column of type character or POSIXct to extract clocktimes from 
to convert to decimal values}
}
\description{
Convert clock time to decimal equivalent
}
\examples{

timedf <- dplyr::tibble(ID = c(1, 1, 2),
DATETIME = c("2022-01-23T12:34:00", "2022-01-28T04:32:00", "2022-01-21T18:00:00"))

times <- lubridate::ymd_hms(timedf$DATETIME)
numeric_time(times)

}
\author{
Samuel P Callisto, PhD
}
