% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_left.R
\name{pad_left}
\alias{pad_left}
\title{add left padding to a vector of values}
\usage{
pad_left(.x, .n, padding_char = "0")
}
\arguments{
\item{.x}{vector}

\item{.n}{total number of characters result should have}

\item{padding_char}{padding char to use, defaults to 0}
}
\description{
add left padding to a vector of values
}
\examples{
pad_left(1, 3)
pad_left(c(1, 10, 100), 4)
}
