% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parens.R
\name{parens}
\alias{parens}
\title{Wrap Text in Parentheses}
\usage{
parens(x, ...)
}
\arguments{
\item{x}{character}

\item{\dots}{ignored}
}
\description{
Wrap text in parentheses.  Useful when formating latex tables.
}
\details{
Text is wrapped in parentheses, with no intermediate spaces.
}
\examples{
parens('x')

}
\author{
Tim Bergsma
}
