% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{peek at the results in a dplyr pipeline}
\usage{
peek(df, n = 5, message = NULL)
}
\arguments{
\item{df}{dataframe in pipeline}

\item{n}{number of rows to show}
}
\description{
peek at the results in a dplyr pipeline
}
\details{
A wrapper around giving a \code{head} command but only as a side effect
so the original data frame is passed along to continue on for further
manipulation in the pipeline
}
\examples{
library(dplyr)
Theoph \%>\% select(Subject, Time, conc) \%>\% peek \%>\% group_by(Subject) \%>\%
summarize(cmax = max(conc))

Theoph \%>\% select(Subject, Time, conc) \%>\% peek(message = "after select") \%>\%
 group_by(Subject) \%>\%
summarize(cmax = max(conc))
}

