% Generated by roxygen2 (4.0.2): do not edit by hand
\name{phx_report}
\alias{phx_report}
\title{create basic folder structure for a phoenix report}
\usage{
phx_report(project, secondary_folders = list("Rscripts", "phoenix-output",
  "for-phoenix", "lab-notebook", "data"))
}
\arguments{
\item{project}{name of top level project folder}

\item{secondary_folders}{list of subfolders}
}
\description{
create basic folder structure for a phoenix report
}
\details{
A placeholder.txt file is created in each subfolder to allow for
initial folder structure to be maintained if using a version control
system such as git (as empty directories are not kept)

Currently only works if there is no project folder with the same name, and
does not do any checking (will just error out)
}
\examples{
\dontrun{
phx_report("Assignment-1")
phx_report("Assignment-1", list("Rscripts", "simulation-results", "data"))
}
}

