% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\name{pool}
\alias{pool}
\title{Test Set Relations Among Two Vectors}
\usage{
pool(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\description{
Test set relations among two vectors.
}
\details{
The two vectors are tested for elements unique to x, unique to y, 
and common to both.
}
\note{
If length of x is zero, it is returned unmodified. If length of 
rule is zero, value is all NA.
}
\examples{
pool(letters[1:5], letters[4:8])

}
\author{
Tim Bergsma
}
