% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_df.R
\name{pool_df}
\alias{pool_df}
\title{Display overlapping names between data.frames}
\usage{
pool_df(x, y)
}
\arguments{
\item{x}{data.frame to be compared}

\item{y}{data.frame to be compared}
}
\description{
List out unique and shared names between data.frames x and y
}
\examples{
library(dplyr)

Theoph_mod <- Theoph \%>\% dplyr::mutate(NEWCOL = 1) \%>\% dplyr::select(-Dose)
pool_df(x = Theoph, y = Theoph_mod)

}
\author{
Samuel P Callisto, PhD
}
