% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_dose.R
\name{pretty_dose}
\alias{pretty_dose}
\title{Overwrite DOSE column into factor and ordered by value}
\usage{
pretty_dose(.dat, .dosecol = "DOSE", .units = "mg")
}
\arguments{
\item{.dat}{data.frame with column `.dosecol`}

\item{.dosecol}{quoted name of column containing dosing information}

\item{.units}{unit to be appended to output dose column}
}
\description{
Overwrite DOSE column into factor and ordered by value
}
\examples{

dose_fix <- pretty_dose(Theoph, .dosecol = "Dose", .units = "mg")
dose_fix$Dose

}
\author{
Samuel P Callisto, PhD
}
