% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_list.R, R/print_plots.R
\name{print_plots}
\alias{print_plots}
\title{create a list of plots cleanly with extra pdf functionality}
\usage{
print_plots(.ggplot_list, .start_page_number = 1)

print_plots(.ggplot_list, .start_page_number = 1)
}
\arguments{
\item{.ggplot_list}{list of ggplot plots}

\item{.start_page_number}{pdf-only starting page number for plots}

\item{.ggplot_list}{list of ggplot plots}

\item{.start_page_number}{pdf-only starting page number for plots}
}
\description{
create a list of plots cleanly with extra pdf functionality

create a list of plots cleanly with extra pdf functionality
}
\details{
Especially for pdf, this can allow the generation of clean pdf pages
with only plots, no code, warnings, etc. for all pages related to the plots. 
In addition, by controlling the start number, you can further trim the pdf
to slice out the extra pages generated from the output but keep a nicely
numbered plot appendix

Especially for pdf, this can allow the generation of clean pdf pages
with only plots, no code, warnings, etc. for all pages related to the plots. 
In addition, by controlling the start number, you can further trim the pdf
to slice out the extra pages generated from the output but keep a nicely
numbered plot appendix
}

