% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{proj_path}
\alias{proj_path}
\title{Calculate the path relative to the project root}
\usage{
proj_path(path)
}
\arguments{
\item{path}{the path to a file or folder.}
}
\description{
Calculate the path relative to the project root
}
