% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{proj_rel}
\alias{proj_rel}
\title{Calculate the path relative to the project root}
\usage{
proj_rel(path)
}
\arguments{
\item{path}{the path to a file or folder.}
}
\description{
Calculate the path relative to the project root
}
