% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_repo.R
\name{pull_repo}
\alias{pull_repo}
\title{function to sparse pull repo}
\usage{
pull_repo(
  .repo,
  .name = NA_character_,
  .dirs = c("deliv", "doc", "renv", "script"),
  .user = "samc"
)
}
\arguments{
\item{.repo}{SVN repo location}

\item{.name}{optional name to assign folder on local disk}

\item{.dirs}{directories to import with depth infinity}

\item{.user}{mc1 username}
}
\description{
function to sparse pull repo
}
