% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sas_folder.R
\name{read_sas_folder}
\alias{read_sas_folder}
\title{Function to read in all SAS files in a given directory (developed by Eric Anderson)}
\usage{
read_sas_folder(.dir)
}
\arguments{
\item{.dir}{directory containing SAS files}
}
\description{
Function to read in all SAS files in a given directory (developed by Eric Anderson)
}
