% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replace_flags.R
\name{replace_char_flags}
\alias{replace_char_flags}
\title{replace symbols or other character flags}
\usage{
replace_char_flags(x, flag, nonflag = NULL, as_numeric = TRUE)
}
\arguments{
\item{x}{vector to replace}

\item{flag}{dataframe with first column being the flag and second the replacement value}

\item{nonflag}{what to do with non-flagged elements}
}
\description{
replace symbols or other character flags
}

