% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_empty.R
\name{replace_empty}
\alias{replace_empty}
\title{Convert empty strings into missing values.}
\usage{
replace_empty(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector with empty strings replaced by missing values.
}
\description{
Convert empty strings into missing values.
}
\examples{
x <- c("a", "", "c")
replace_empty(x)
}
