% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_from_template.R
\name{replace_from_template}
\alias{replace_from_template}
\title{replace values from a template file}
\usage{
replace_from_template(file, patterns, replacement_df, output_dir = NULL,
  file_name = NULL, ...)
}
\arguments{
\item{file}{template file}

\item{patterns}{patterns to match}

\item{replacement_df}{dataframe of replacements}

\item{output_dir}{set directory if not current directory}

\item{file_name}{column in replacement containing requested output file name}

\item{...}{additional arguments for readLines}
}
\description{
replace values from a template file
}
\details{
the pattern will first check for names, and if matching names in replacement df will follow, however
if no names are detected, then will match by position
}

