% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_list_elements.R
\name{replace_list_elements}
\alias{replace_list_elements}
\title{replace list elements by name}
\usage{
replace_list_elements(.list, replacement, add = TRUE)
}
\arguments{
\item{.list}{original list}

\item{replacement}{replacement list}

\item{add}{add replacement values if they don't exist, default = TRUE}
}
\description{
replace list elements by name
}
\details{
Finds and replaces list elements by name and throws an error if an 
   element is not available in the original list.
}
\examples{
olist <- list(dv = "dv", idv = "idv", notreplace = "somethingelse")
replacement <- list(dv = "conc")
combined_list <- replace_list_elements(olist, replacement)
}
