% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replace_flags.R
\name{replace_values}
\alias{replace_values}
\title{replace symbols or other character flags}
\usage{
replace_values(x, flag_df, nonflag = NULL, as_numeric = TRUE)
}
\arguments{
\item{x}{vector to replace}

\item{flag_df}{dataframe with first column being the flag and second the replacement value}

\item{nonflag}{what to do with non-flagged elements}

\item{as_numeric}{whether to return column as numeric}
}
\description{
replace symbols or other character flags
}
\examples{
df <- data.frame(ID = 1, DV = c(1, "BQL", ".", 5))
rflags <- data.frame(flag = c("BQL", "."), replacement = -99)
df$DVR <- replace_values(df$DV, rflags)
}

