% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right.R
\name{right}
\alias{right}
\title{Extract letters from end of string}
\usage{
right(.word, .length)
}
\arguments{
\item{.word}{string to extract characters from}

\item{.length}{number of letters to extract from end of .word}
}
\description{
Based on the RIGHT() function from Microsoft Excel
}
\examples{

# If you want to get the end 3 characters from a string
right("string-123", 3)

}
\author{
Samuel P Callisto, PhD
}
