% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rm_tables}
\alias{rm_tables}
\title{clean tables associated with nonmem output from directory}
\usage{
rm_tables(runno, xpose_tables = c("sdtab", "mutab", "patab", "catab", "cotab",
  "mytab", "extra", "xptab", "cwtab"), table_names = NULL)
}
\arguments{
\item{runno}{nonmem run number to check for tables}

\item{xpose_tables}{}

\item{table_names}{additional table names}
}
\description{
clean tables associated with nonmem output from directory
}
\details{
defaults to check and clean for xpose-style tables
}

