% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_compare.R
\name{row_compare}
\alias{row_compare}
\title{Find all columns where rows are not identical}
\usage{
row_compare(.df)
}
\arguments{
\item{.df}{data.frame containing rows to compare}
}
\description{
Return a data.frame containing only the columns where the rows
are not identical.
}
\examples{
row_compare(mtcars[1:3,])

}
