% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s_pauc.R
\name{s_pauc}
\alias{s_pauc}
\title{summarize paucs}
\usage{
s_pauc(df, time, dv, paucs, digits = Inf)
}
\arguments{
\item{df}{data frame}

\item{time}{string name for time column for pauc slice}

\item{dv}{string name for dependent variable column (eg. dv or cobs)}

\item{paucs}{list of ranges for pauc calculation}
}
\description{
summarize paucs
}
\examples{
\dontrun{
library(PKPDdatasets)
sd_oral_richpk  \%>\% group_by(ID) \%>\% s_pauc
sd_oral_richpk  \%>\% group_by(Gender) \%>\% s_quantiles("Conc", probs = c(0.1, 0.5))
}
}

