% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s_pauc.R
\name{s_pauc_i}
\alias{s_pauc_i}
\title{summarize quantile}
\usage{
s_pauc_i(df, time, dv, range, digits = Inf)
}
\arguments{
\item{df}{data frame}

\item{time}{string name for time column for pauc slice}

\item{dv}{string name for dependent variable column (eg. dv or cobs)}

\item{range}{whether to remove na values}

\item{digits}{number of digits to pass to round}
}
\description{
summarize quantile
}
\details{
for internal use in the s_pauc function
}

