% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R
\name{s_quantile}
\alias{s_quantile}
\title{summarize quantile}
\usage{
s_quantile(df, col_name, prob, na.rm = T)
}
\arguments{
\item{df}{data frame}

\item{col_name}{string name for column to calculate quantiles on}

\item{prob}{probability for quantiles of interest}

\item{na.rm}{whether to remove na values}
}
\description{
summarize quantile
}
