% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-df.R
\name{search_df}
\alias{search_df}
\title{Search for value in a data.frame}
\usage{
search_df(.df, .value)
}
\arguments{
\item{.df}{a data.frame for the count of a value in each column}

\item{.value}{numeric or character value to search}
}
\description{
\code{search_df} counts the number of occurrences of a specified value in 
each column of a given data.frame. The output will show the count for each
column that had at least one value matching the search criteria.
}
\examples{
search_df(Theoph, 1)

}
