% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_bins_cpp}
\alias{set_bins_cpp}
\title{given a set of bin ranges, assign each value to a bin}
\usage{
set_bins_cpp(x, left, right)
}
\arguments{
\item{x}{A numeric vector of values}

\item{left, right}{Boundary values}
}
\description{
given a set of bin ranges, assign each value to a bin
}
\details{
Given a set of quantiles/bins/etc established from a separate dataset, it can 
be useful to assign the same bins to new or simulated data for comparisons
or to do additional analysis such as assign dropouts etc. This function can be
used to take the breakpoints to establish bins quickly and easily
}
\keyword{internal}
