% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_labels.R
\name{show_labels}
\alias{show_labels}
\title{Show column labels}
\usage{
show_labels(df, ...)
}
\arguments{
\item{df}{data frame}

\item{...}{columns of interest. If none are specified then all columns will
be used. Accepts:
\itemize{
 \item unquoted column names (e.g., \code{ID})
 \item character column names (e.g., \code{"ID"})
 \item numeric indices (e.g., \code{1:3})
 \item tidyselect helpers (e.g., \code{starts_with("A")})
 \item vectors of any of the above
}}
}
\value{
A tibble with the names of \code{df} in the first column and the
  labels of \code{df} in the second column
}
\description{
Gives column names and labels
}
\details{
If a column has a vector of labels then the first label will be returned.
}
\examples{
id <- 1:4
age <- c(83, 29, 64, 40)
bwt <- c(71, 66, 70, 81)
attr(id, "label") <- "Subject ID"
attr(age, "label") <- "Age (years)"
attr(bwt, "label") <- "Weight (kg)"
df <- tibble::tibble(ID = id, AGE = age, BWT = bwt)
show_labels(df)
show_labels(df, "AGE", BWT)
show_labels(df, 2)
show_labels(df, c("BWT", "ID"))
show_labels(df, tidyselect::starts_with("A"))
}
\seealso{
\link[tidyselect]{starts_with}
}
