% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{tf_options}
\alias{tf_options}
\alias{tf_options_clear}
\alias{mrg_script}
\alias{tables_to}
\alias{figures_to}
\title{Set options for annotating and saving tables and figures}
\usage{
tf_options()

tf_options_clear(quietly = FALSE)

mrg_script(path = NULL)

tables_to(path, set_script = TRUE, path.type = "proj")

figures_to(path, proj_path = NULL, set_script = TRUE)
}
\arguments{
\item{quietly}{if \code{TRUE}, don't echo \code{tf_options()} outputs after options
are cleared with \code{tf_options_clear()}.}

\item{path}{the absolute path or path relative to the current working
directory.}

\item{set_script}{if \code{TRUE} (the default), include a call to \code{mrg_script()}
to set the \code{mrg.script} option.}

\item{path.type}{indicates how \code{pmtables} should format the path portion
of table annotations; see \code{pmtables::format_table_path}.}
}
\value{
\itemize{
\item \code{tables_to()}, \code{figures_to()} and \code{mrg_script()} return the value of the
option invisibly
\item \code{tf_options()} prints messages to the console and returns \code{NULL} invisibly
\item \code{tf_options_clear()} prints \code{tf_options()} output unless the user request
quiet reset.
}
}
\description{
\itemize{
\item \code{mrg_script()} sets the \code{mrg.script} option using \code{\link[=this_file_proj]{this_file_proj()}}.
\item \code{tables_to()} sets the \code{pmtables.dir} option as the default table output
directory; this function also sets the \code{pmtables.path.type} option,
defaulting to \code{proj}.
\item \code{figures_to()} sets the \code{mrggsave.dir} option as the default figure output
directory.
\item \code{tf_options()} prints the \code{mrg.script}, \code{pmtables.dir}, and \code{mrggsave.dir}
options to the console.
\item \code{tf_options_clear()} clears options set by \code{mrg_script()}, \code{tables_to()},
and \code{figures_to()}, setting the option value to \code{NULL}.
}
}
