% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{this_file_here}
\alias{this_file_here}
\title{Return the relative path to the current script}
\usage{
this_file_here()
}
\value{
The name and path to the current script, relative to a root
directory defined by a call to \code{\link[here:here]{here::here()}}.
}
\description{
Return the relative path to the current script
}
\seealso{
\code{\link[=this_file_path]{this_file_path()}}, \code{\link[=this_file_name]{this_file_name()}}
}
