% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{this_file_name}
\alias{this_file_name}
\title{Return the name of the current script}
\usage{
this_file_name()
}
\value{
The name of the current script, with no path information.
}
\description{
Return the name of the current script
}
\seealso{
\code{\link[=this_file_path]{this_file_path()}}, \code{\link[=this_file_project]{this_file_project()}}
}
