% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{this_file_path}
\alias{this_file_path}
\title{Return the absolute path to the current script}
\usage{
this_file_path()
}
\value{
The absolute path to the current script.
}
\description{
Return the absolute path to the current script
}
\seealso{
\code{\link[=this_file_name]{this_file_name()}}, \code{\link[=this_file_here]{this_file_here()}}
}
