% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{this_file_proj}
\alias{this_file_proj}
\title{Return the path to current script relative to its project}
\usage{
this_file_proj()
}
\value{
The file name of the current script, relative to the root of the project that
contains it, as defined by a call to \code{\link[this.path:here]{this.path::this.proj()}}.
}
\description{
Return the path to current script relative to its project
}
\seealso{
\code{\link[=this_file_path]{this_file_path()}}, \code{\link[=this_file_name]{this_file_name()}}
}
