% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wam}
\alias{wam}
\title{wam function}
\usage{
wam(ncovtheta, cov_theta_nums, npar, nobs, run_num, dir = NULL)
}
\arguments{
\item{ncovtheta}{number of covariate thetas}

\item{cov_theta_nums}{<- vector of theta numbers for covariate thetas}

\item{npar}{number of parameters in model (theta, omega, sigma)}

\item{nobs}{number of observations}

\item{dir}{directory where .cov and .ext files located}

\item{runnum}{run number}
}
\description{
wam function
}
\details{
for runnum it is recommended to pass as a string or leading 0's will be stripped
}

