% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_numeric}
\alias{as_numeric}
\title{convert to numeric passing through character for safety}
\usage{
as_numeric(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{additional argument to as.character}
}
\description{
convert to numeric passing through character for safety
}
\examples{
# factor with weird levels that we don't want to keep
ex <- factor(c(1, 2, 3, 4), levels = c(2, 3, 1, 4)) 
ex

# keeps information about the levels, oh no!
as.numeric(ex) 

# keeps the labelled values
as_numeric(ex)
}

