% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric.R
\name{as_numeric}
\alias{as_numeric}
\title{convert to numeric passing through character for safety}
\usage{
as_numeric(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{additional argument to as.character}
}
\description{
convert to numeric passing through character for safety
}

