% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AUC_inf.R
\name{auc_inf}
\alias{AUC_inf}
\alias{auc_inf}
\title{Calculate AUCt-inf}
\usage{
auc_inf(time, conc, last_points = c(3, 4, 5))

AUC_inf(...)
}
\arguments{
\item{time}{column name for time}

\item{conc}{column name for conc}

\item{last_points}{vector of amount of points in terminal phase that will be evaluated for extrapolation}

\item{...}{args to pass to auc_inf}
}
\description{
Calculate AUCt-inf
}
\details{
last_points defaults to 3, 4, 5
see auc_partial for other details
}

