% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AUC_partial.R
\name{AUC_partial}
\alias{AUC_partial}
\title{Calculate partial AUC}
\usage{
AUC_partial(time = "TIME", conc = "DV", partialtime = 28,
  begin_time = 0)
}
\arguments{
\item{time}{column name for time}

\item{conc}{column name for conc}

\item{partialtime}{final time for partial AUC}

\item{begin_time}{starting time for partial AUC}
}
\description{
Calculate partial AUC
}

