\name{AUC_partial}
\alias{AUC_partial}
\title{Calculate partial AUC}
\usage{
AUC_partial(time = "TIME", conc = "DV", partialtime = 28)
}
\arguments{
  \item{time}{column name for time}

  \item{conc}{column name for conc}

  \item{partialtime}{final time for partial AUC}
}
\description{
Calculate partial AUC
}

