% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_theme.R
\name{base_theme}
\alias{base_theme}
\title{A theme with better default values for pharmacometric plots, especially conc-time}
\usage{
base_theme(
  legend_text = 10,
  legend_title = 12,
  axis_title_x = 12,
  axis_title_y = axis_title_x,
  axis_text_x = 10,
  axis_text_y = axis_text_x,
  strip_text_x = 12,
  strip_text_y = strip_text_x
)
}
\arguments{
\item{legend_text}{size of legend text}

\item{legend_title}{legend title}

\item{axis_title_x}{size of X axis title}

\item{axis_title_y}{size of Y axis title}

\item{axis_text_x}{size of X axis text}

\item{axis_text_y}{size of Y axis text}

\item{strip_text_x}{size of strip text for X axis}

\item{strip_text_y}{size of strip text for Y axis}
}
\description{
A theme with better default values for pharmacometric plots, especially conc-time
}
