% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/base_theme.R
\name{base_theme}
\alias{base_theme}
\title{A theme with better default values for pharmacometric plots, especially conc-time}
\usage{
base_theme()
}
\description{
A theme with better default values for pharmacometric plots, especially conc-time
}

