% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bring_to_front.R
\name{bring_to_front}
\alias{bring_to_front}
\title{bring select columns to the front of a dataframe}
\usage{
bring_to_front(.df, ...)
}
\arguments{
\item{.df}{dataframe with column order to adjust}

\item{...}{columns to bring to front}
}
\description{
bring select columns to the front of a dataframe
}
\examples{
head(Theoph)
head(bring_to_front(Theoph, conc, Time))
}
