% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize_names.R
\name{capitalize_names}
\alias{capitalize_names}
\title{Capitalize all names for a dataframe}
\usage{
capitalize_names(df)
}
\arguments{
\item{df}{data frame to capitalize names}
}
\description{
Capitalize all names for a dataframe
}
\details{
This is a simple wrapper function to reduce typing and more
easily pass data as it is read from a file
}
\examples{
names(Theoph)
[1] "Subject" "Wt"      "Dose"    "Time"    "conc" 

cTheoph <- capitalize_names(Theoph)
names(cTheoph)
[1] "SUBJECT" "WT"      "DOSE"    "TIME"    "CONC" 

}
