% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal_chaining.R
\name{\%>\%}
\alias{\%>\%}
\title{internal Chaining}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{The value to be piped.}

\item{rhs}{A function or expression.}
}
\description{
\code{\%>\%} - The \pkg{magrittr} "then" chain operator imported by
\pkg{magrittr}.  Imported for convenience.  See
\url{http://www.rdocumentation.org/packages/magrittr/functions/magrittr}
for details.
}

