% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gradients.R
\name{check_gradients}
\alias{check_gradients}
\title{get gradients for nonmem run running via PSN}
\usage{
check_gradients(runnum, wd = NULL, folder_str = ".mod.dir.1",
  full_directory = NULL, print = FALSE)
}
\arguments{
\item{runnum}{Xpose-style run number}

\item{wd}{relative directly location where PSN run results are output}

\item{folder_str}{folder name besides run number}

\item{full_directory}{can be specified if folder where gradient file stored non-traditional}

\item{print}{print visual plot of gradients}
}
\description{
get gradients for nonmem run running via PSN
}
\details{
by default, the working directory should be set to the location of .mod file associated with the run

the default folder structure anticipates runs were run via psn with nomenclature such as run001.mod -mod

has not been checked for side effects if have multiple runs with the same file name
due to restarts or other changes

default to not print visual of gradients
}
\examples{
\dontrun{
check_gradients(001)
check_gradients(001, wd = "~/ProjX/NONMEM")
check_gradients(001, folder_str = "modelfit_dir1")
check_gradients(001, full_directory = "~/user/ProjX/NONMEM/run001.mod.dir.1/NM_run2/psn.grd")
}
}

