% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_df.R
\name{chunk_df}
\alias{chunk_df}
\title{chunk dataframes so easy to split for parallel processing}
\usage{
chunk_df(.gdf, ..., .nchunks = parallel::detectCores(), .as_list = TRUE)
}
\arguments{
\item{.gdf}{(grouped) data frame}

\item{...}{grouping variables, either a character vector or NSE-style column names}

\item{.nchunks}{set number of chunks}

\item{.as_list}{return df as a list}
}
\description{
chunk dataframes so easy to split for parallel processing
}
\examples{
library(dplyr)
Theoph \%>\% chunk_df()

Theoph \%>\% group_by(Subject) \%>\% chunk_df()
Theoph \%>\% chunk_df(Subject)
Theoph \%>\% chunk_df(c("Subject"))

Theoph \%>\% chunk_df(Subject, .nchunks = 3)

# to retain the df without being a list
Theoph \%>\% chunk_df(Subject, .nchunks = 3, .as_list = FALSE)
}
