% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-subject-check.R
\name{df_subject_check}
\alias{df_subject_check}
\title{Subject data checks}
\usage{
df_subject_check(
  .df,
  .subject_col = NULL,
  .dose_col = "DOSE",
  .evid_col = "EVID",
  .blq_col = "BLQ"
)
}
\arguments{
\item{.df}{Data.frame to perform checks on}

\item{.subject_col}{Character column name of subject identifier. (default to ID or USUBJID)}

\item{.dose_col}{Character column name of dose level. (default to DOSE)}

\item{.evid_col}{Character column name of event indicator variable. (default to EVID)}

\item{.blq_col}{Character column name of below limit of quantification indicator (optional)}
}
\description{
Use this function to perform quick checks on a data.frame such as:
- One row per subject
- One dose level per subject
- Multiple EVID's per subject
}
