% Generated by roxygen2 (4.0.1): do not edit by hand
\name{file_rows}
\alias{file_rows}
\title{count number of rows in a file without needing to read file into R}
\usage{
file_rows(file)
}
\arguments{
\item{file}{file path to be read in}
}
\description{
count number of rows in a file without needing to read file into R
}
\details{
uses wc -l system call to count rows
}

