% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_functions.R
\name{get_key}
\alias{get_key}
\title{find unique values for key}
\usage{
get_key(df, key_cols = names(df))
}
\arguments{
\item{df}{data frame}

\item{key_cols}{vector of column names. Defaults to all columns}
}
\description{
find unique values for key
}
\details{
used in stratify_df, however also made available to find
non-duplicated values across multiple columns 
when !duplicated(col) is not sufficient
}

