% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_life.R
\name{half_life}
\alias{half_life}
\title{Estimate elimination half-life from 2-compartment model in hours}
\usage{
half_life(cl, vc, vp, q)
}
\arguments{
\item{cl}{clearance estimate (L/hr)}

\item{vc}{central volume estimate (L)}

\item{vp}{peripheral volume estimate (L)}

\item{q}{intercompartmental clearance estimate (L/hr)}
}
\description{
Estimate elimination half-life from 2-compartment model in hours
}
