% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ids_per_plot.R
\name{ids_per_plot}
\alias{ids_per_plot}
\title{split IDs into groups to use for subsequent plotting}
\usage{
ids_per_plot(id, id_per_plot = 9)
}
\arguments{
\item{id}{vector of ids (eg id column)}

\item{id_per_plot}{number of ids per plot. Default to 9}
}
\description{
split IDs into groups to use for subsequent plotting
}
\details{
works very well with hadley wickham's lowliner package to create a column
to split on then subsequently plot
}
\examples{
\dontrun{
library(PKPDdatasets)
sd_oral_richpk$IDPP <- ids_per_plot(sd_oral_richpk$ID)
library(dplyr)
sd_oral_richpk <- sd_oral_richpk \%>\% mutate(IDPP = ids_per_plot(ID, 16))
library(lowliner)
sid <- sd_oral_richpk \%>\% group_by(ID) \%>\% split(.[["IDPP"]])

library(ggplot2)
p_split_id <- function(df) {
  p <- ggplot(df, aes(x = Time, y = Conc)) +
    geom_point() + geom_line() + facet_wrap(~ID) + base_theme()
  suppressWarnings(print(p))
  return(p)
}

sid \%>\% map(p_split_id)
}
}

