% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_through}
\alias{max_through}
\title{give the max value up to that point}
\usage{
max_through(x)
}
\arguments{
\item{x}{A numeric vector of values}
}
\description{
give the max value up to that point
}
\details{
useful for safety analyses where an event may be defined as a certain change
in a biomarker, so need to see how the current measurement compares to the maximum
value up to that point
}
\examples{
max_through(c(4, 3, 3, 2, 5, 1))
max_through(c(NA, 2, 1, 4, 2))
}

