% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{min_through}
\alias{min_through}
\title{give the min value up to that point}
\usage{
min_through(x)
}
\arguments{
\item{x}{A numeric vector of values}
}
\description{
give the min value up to that point
}
\details{
useful for safety analyses where an event may be defined as a certain change
in a biomarker, so need to see how the current measurement compares to the minimum
value up to that point
}
\examples{
/dontrun {
min_through(c(4, 3, 3, 2, 4, 1))
min_through(c(NA, 2))
}
}

