% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasum.R
\name{nasum}
\alias{nasum}
\title{Count Number of \code{NA} Values in List Items}
\usage{
nasum(x,simplify=TRUE)
}
\arguments{
\item{x}{list of vector-like objects; e.g. a data frame}

\item{simplify}{logical: coerce to vector result if possible}
}
\description{
\code{nasum} Count the number of \code{NA} values in each element of a list-like object.
}
\details{
Arguments are passed to \code{simplify}, along with a function 
that sums instances of \code{NA}.
}
\examples{
nasum(Theoph)

}
\author{
Natalie Hsiang
}
