% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA.R
\name{nca}
\alias{nca}
\alias{NCA}
\title{calculate basic NCA parameters t1/2 Cl, Cmax etc}
\usage{
nca(.time, dv, dose, last_times = c(3, 4, 5), digits = 2)

NCA(...)
}
\arguments{
\item{.time}{vector of times}

\item{dv}{vector of observations (concentrations)}

\item{dose}{vector or single value of Dose given}

\item{last_times}{vector of numbers of time points to evaluate for 
AUCinf extrapolation default 3-5}

\item{digits}{number of digits to round to. Can use NULL for no rounding}

\item{...}{arguments to pass to nca}
}
\description{
calculate basic NCA parameters t1/2 Cl, Cmax etc
}
