% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nm_description}
\alias{nm_description}
\title{show description output from .mod file}
\usage{
nm_description(filename)
}
\arguments{
\item{filename}{name of .mod file}
}
\value{
string
}
\description{
show description output from .mod file
}
\details{
return line containing description file
}
\seealso{
nm_notes
}

