% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nm_comments.R
\name{nm_notes}
\alias{nm_notes}
\title{show notes output from .mod file}
\usage{
nm_notes(filename)
}
\arguments{
\item{filename}{name of .mod file}
}
\value{
string
}
\description{
show notes output from .mod file
}
\details{
returns all lines between ;NOTES and ;ENDNOTES in .mod file
if filename not in current wd, need to give path as well
}
\examples{
\dontrun{
nm_notes("run102.mod")
}
}
\seealso{
nm_description
}

