% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsub.R
\name{nsub}
\alias{nsub}
\title{Count IDs}
\usage{
nsub(.df, .id = USUBJID)
}
\arguments{
\item{.df}{data.frame containing column .id}

\item{.id}{column name that will be used to uniquely identify IDs}
}
\description{
Count IDs
}
