% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsub.R
\name{nsub}
\alias{nsub}
\title{Count number of IDs in a NONMEM data set}
\usage{
nsub(.df, .id)
}
\arguments{
\item{.df}{data.frame containing column .id}

\item{.id}{column name that will be used to uniquely identify IDs}
}
\description{
Count number of IDs in a NONMEM data set
}
\examples{

nsub(.df = Theoph, .id = Subject)

}
\author{
Samuel P Callisto, PhD
}
