% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_time.R
\name{numeric_time}
\alias{numeric_time}
\title{Convert clock time to decimal equivalent}
\usage{
numeric_time(.col)
}
\arguments{
\item{.col}{column of type character or POSIXct to extract clocktimes from 
to convert to decimal values}
}
\description{
Convert clock time to decimal equivalent
}
