% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_left.R
\name{pad_left}
\alias{pad_left}
\title{Add character to left of every character}
\usage{
pad_left(.x, .n, padding_char = "0")
}
\arguments{
\item{.x}{vector}

\item{.n}{total number of characters result should have}

\item{padding_char}{padding char to use, defaults to 0}
}
\description{
Add character to left of every character
}
\details{
Use this function to add specific character to the
left of each element in a given vector.

The output is a character vector and default padding
character is "0".
}
\examples{
pad_left(1, 3)
[1] "001"

pad_left(c(1, 10), 4, "Z")
[1] "ZZZ1" "ZZ10"

}
