% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_dose.R
\name{pretty_dose}
\alias{pretty_dose}
\title{function to overwrite DOSE column into factor and ordered by value}
\usage{
pretty_dose(.dat, .dosecol = "DOSE", .units = "mg")
}
\arguments{
\item{.dat}{data.frame with column `.dosecol`}

\item{.dosecol}{quoted name of column containing dosing information}

\item{.units}{unit to be appended to output dose column}
}
\description{
function to overwrite DOSE column into factor and ordered by value
}
