% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_csv_wunits}
\alias{read_csv_wunits}
\title{to more easily read data with a 2nd row with units}
\usage{
read_csv_wunits(data, skip_top = 0, stringsAsFactors = FALSE, ...)
}
\arguments{
\item{data}{the name of the csv}

\item{skip_top}{rows to skip before header row}

\item{stringsAsFactors}{whether to include string columns as factors}

\item{...}{additional arguments to read.csv functions}
}
\value{
data frame of read-in csv
}
\description{
to more easily read data with a 2nd row with units
}
\details{
helpful function to handle situations where the second row is a units row
as often seen with phoenix-style datasets
}
\examples{
\dontrun{
read_csv_wunits("example.csv")
read_csv_wunits("example.csv", skip_top = 1) # will ignore 1st line, good for comment lines
}
}

