% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_nonmem}
\alias{read_nonmem}
\title{Better defaults for reading in nonmem data}
\usage{
read_nonmem(file, skip = 1, header = TRUE, colClasses = "numeric", ...)
}
\arguments{
\item{file}{file to be read in}

\item{skip}{number of lines to skip (default = 1)}

\item{header}{TRUE/FALSE header}

\item{colClasses}{numeric default to colClasses argument in read.table}

\item{...}{dots argument to pass to read.table}
}
\value{
data frame object
}
\description{
Better defaults for reading in nonmem data
}
\details{
gives colClasses=numeric + default args to read.table
}

