% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_xpdb.R
\name{read_xpdb}
\alias{read_xpdb}
\title{read xpose tables to xpose database and save and compress as RDS}
\usage{
read_xpdb(runno, table.names = c("sdtab", "mutab", "patab", "catab", "cotab",
  "mytab", "extra", "xptab", "cwtab"), ..., RDS_name = paste0("xpdb", runno))
}
\arguments{
\item{runno}{}

\item{...}{arguments to pass to xpose.data}

\item{rm_tabs}{logical value to delete output tables after saving to xpdb}
}
\description{
read xpose tables to xpose database and save and compress as RDS
}
\examples{
\dontrun{
read_xpdb("001") # would save a file xpdb001.rds
}
}

