% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_dots.R
\name{replace_dots}
\alias{replace_dots}
\title{Convert '.' values into missing values.}
\usage{
replace_dots(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector with '.' replaced by NA values.
}
\description{
Convert '.' values into missing values.
}
\details{
good for nonmem preprocessing as missing values are represented with '.' in NONMEM
}
\examples{
x <- c(".", "1", "1")
replace_dots(x)
}
