% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replace_list_elements.R
\name{replace_list_elements}
\alias{replace_list_elements}
\title{replace list elements by name}
\usage{
replace_list_elements(list, replacement)
}
\arguments{
\item{list}{original list}

\item{replacement}{replacement list}
}
\description{
replace list elements by name
}
\details{
Finds and replaces list elements by name and throws an error if an
   element is not available in the original list.
}
\examples{
\dontrun{
list <- list(dv = "dv", idv = "idv")
replacement <- list(dv = "conc", idv = "time")
list <- replace_list_elements(list, replacement)
}
}

