% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resample_df}
\alias{resample_df}
\title{resampling}
\usage{
resample_df(df, key_cols, strat_cols = NULL, key_col_name = "KEY")
}
\arguments{
\item{df}{}

\item{key_cols}{}

\item{strat_cols}{}

\item{key_name}{name of key column}

\item{...}{dots}
}
\description{
resampling
}

