% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resampling_functions.R
\name{resample_df}
\alias{resample_df}
\title{resampling}
\usage{
resample_df(df, key_cols, strat_cols = NULL, key_col_name = "KEY",
  n = NULL)
}
\arguments{
\item{df}{}

\item{key_cols}{}

\item{strat_cols}{}

\item{key_col_name}{name of key column}

\item{n}{number of unique sampled keys, defaults to match dataset}
}
\description{
resampling
}

