% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_quantiles.R
\name{s_quantiles_}
\alias{s_quantiles_}
\alias{s_quantiles}
\title{summarize quantiles for a column}
\usage{
s_quantiles_(.data, x, probs, na_rm = TRUE)

s_quantiles(.data, x, probs, na_rm = TRUE)
}
\arguments{
\item{.data}{data frame}

\item{x}{column to calculate quantiles for}

\item{probs}{probabilities to calculate quantiles for}

\item{na_rm}{remove na's before calculating value for quantile}
}
\description{
summarize quantiles for a column
}
\examples{
library(dplyr)
sd_oral_richpk \%>\% group_by(Gender, Time) \%>\% s_quantiles(Conc, c(0.05, 0.5, 0.95))
}
