% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quantiles.R
\name{s_quantiles}
\alias{s_quantiles}
\title{summarize quantiles}
\usage{
s_quantiles(df, col_name, probs, na.rm = T)
}
\arguments{
\item{df}{data frame}

\item{col_name}{string name for column to calculate quantiles on}

\item{probs}{probability for quantiles of interest}

\item{na.rm}{whether to remove na values}
}
\description{
summarize quantiles
}
\details{
can calculate multiple quantiles by passing in a vector of probabilities.
Can also pass in a grouped df using dplyr::group_by to get summaries by group
}
\examples{
\dontrun{
library(PKPDdatasets)
sd_oral_richpk  \%>\% s_quantiles("Conc", probs = c(0.1, 0.5))
sd_oral_richpk  \%>\% group_by(Gender) \%>\% s_quantiles("Conc", probs = c(0.1, 0.5))
}
}

