% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantiles.R, R/s_quantiles.R
\name{s_quantiles_}
\alias{s_quantiles}
\alias{s_quantiles_}
\title{summarize quantiles}
\usage{
s_quantiles_(.data, x, probs, na_rm = TRUE)

s_quantiles(.data, x, probs, na_rm = TRUE)

s_quantiles_(.data, x, probs, na_rm = TRUE)

s_quantiles(.data, x, probs, na_rm = TRUE)
}
\arguments{
\item{.data}{data frame}

\item{x}{column to calculate quantiles for}

\item{probs}{probability for quantiles of interest}

\item{na_rm}{remove na's before calculating value for quantile}

\item{df}{data frame}

\item{col_name}{string name for column to calculate quantiles on}

\item{na.rm}{whether to remove na values}

\item{quantiles}{probabilities to calculate quantiles for}
}
\description{
summarize quantiles

summarize quantiles for a column
}
\details{
can calculate multiple quantiles by passing in a vector of probabilities.
Can also pass in a grouped df using dplyr::group_by to get summaries by group
}
\examples{
\dontrun{
library(PKPDdatasets)
sd_oral_richpk  \%>\% s_quantiles("Conc", probs = c(0.1, 0.5)) 
sd_oral_richpk  \%>\% group_by(Gender) \%>\% s_quantiles("Conc", probs = c(0.1, 0.5)) 
}
\dontrun{
library(dplyr)
library(PKPDdatasets)
sd_oral_richpk \%>\% group_by(Gender, Time) \%>\% s_quantiles(Conc, c(0.05, 0.5, 0.95))
}
}

