% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set_bins.R
\name{set_bins}
\alias{set_bins}
\title{given a set of bin ranges, assign each value to a bin}
\usage{
set_bins(x, breaks = quantile(x), lower_bound = -Inf, upper_bound = Inf,
  quiet = TRUE)
}
\arguments{
\item{x}{numeric vector to assign bins}

\item{breaks}{breaks for each bin, defaults to quantiles}

\item{lower_bound}{set a lower bound for the first bin, defaults to -Inf}

\item{upper_bound}{set an upper bound for the last bind, defaults to Inf}

\item{quiet}{whether to give additonal information regarding bins and assigned range for each}
}
\description{
given a set of bin ranges, assign each value to a bin
}
\details{
Given a set of quantiles/bins/etc established from a separate dataset, it can
be useful to assign the same bins to new or simulated data for comparisons
or to do additional analysis such as assign dropouts etc. This function can be
used to take the breakpoints to establish bins quickly and easily

If there is concern over data being outside the range of the assigned breaks,
one can assign -Inf to lower and/or Inf to upper to make sure all values will be
assigned to a bin
}

