% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_functions.R
\name{stratify_df}
\alias{stratify_df}
\title{stratify based on some columns}
\usage{
stratify_df(df, strat_cols, n, replace = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{strat_cols}{columns to stratify on}

\item{n}{number of samples}

\item{replace}{whether to resample with replacement}
}
\description{
stratify based on some columns
}
\keyword{internal}
